requirements.txt
streamlit>=1.36
pandas>=2.0
sqlalchemy>=2.0
psycopg2-binary
plotly>=5.20
python-dotenv
